Add slashing instructions to finalized blocks.
This adds an `Instruction` type, and a field to `FinalizedBlock`, where
the consensus protocol can communicate slashing and reward instructions.

Rewards are not being computed yet, but slashing is.

This PR contains only the consensus component side, i.e. creation of the
instructions. Handling them will need to be implemented in the
`BlockExecutor`.