Fix era changes, parent block and test.
The first `FinalizedBlock` in a new era must have the previous era's
switch block as a parent, not `None`.

Consensus protocol results from starting a new era must have the _new_
era's ID, not the old ones.

Eras start at 0, so if we are in era 2, we will have three entries: The
test checked only for two.