NDRS-311: remove endpoint type from small_network
This commit lays the groundwork for CEP-0010.  It sees the replacement
of endpoints by SocketAddrs for incoming connections, and the removal of
the retry mechanism for when a connection fails.

It also removes the option to use pre-generated certificates in the
network component, and removes the option to create certificates at the
CLI.