Remove round exponent validity conditions.
The only thing that's important is that the round exponent doesn't
change within a round.

Limiting how fast it can change and how far it can diverge from the
median was only meant to protect against spam attacks, but setting a
minimal round exponent is a much easier and effective way to do that.