Adapt test to end-of-era rewards, fix bug.
The `compute_rewards_test` was updated: Rewards are always applied to
all ancestors of the payout block, which in production will be the
switch block.

To make the test easier to read, the rewards for individual ancestors
are also verified.

Quora in reward calculation apply only to level-1 summits, and don't
include the faulty validators. Such a summit finalizes the block if and
only if it has > 50% of the weight, regardless of the already known
faulty validators. This was incorrect in `compute_rewards_for`.