EE-1062: Pay for storage writes.
This is done on the write end - all access to tracking copy is guarded
and never exposed directly from within executing context. This way the
only way to write anything is through newly introduced `metered_write`
method which checks first if a given amount for storage bytes can be
charged for and then issues the write.

Keep in mind that writes still go through tracking copy as in - they are
cached first, but this is considered fine.