EE-1082: Put round seigniorage rate in chainspec.
This also changes the formula which I believe was wrong so far: based on
the rustdoc I assume it was calculated on floats in python which as
limited precision. New fraction was computed using wolfram alpha where
original formula from the rustdoc was factorized which resulted in
current fraction of 6414/623437335209.