Call announce_proposed_proto_block later.
It's wrong to call it when the proto block has been validated, because
at that point it's not necessarily passed into the consensus protocol
yet: It may still be waiting for evidence that it depends on.

Only once it's passed in will the protocol be able to detect
finalization (or later maybe orphans).