NO-TICKET: Clean up temporary directory.
A lazy_static variables just does not ever call destructors - therefore
any temporary directory which a Drop impl that does cleanup is never
ran. This commit changes the behavior of such tests so a temporary
directory is created on each tests that requires it, and then manually
removes it at the end of tests. If ever such test fail, the data is not
removed giving you an option to inspect the filesystem.