Delay era deactivation.
Since summits are not monotonous, it can happen that we see a switch
block finalized but other honest validators don't, because they first see
an equivocation that invalidates the summit.

To make it more likely that everyone sees the switch block finalized, we
keep voting for a while, and only stop once we observed no new
equivocations within a period that was sufficient to finalize the block.