Pause if too few validators are online.
This adds a ping mechanism that allows validators to signal that they
are online even if they are not creating new units. If we see fewer
validators online than the required quorum for finality at the
configured fault tolerance threshold, we pause and don't create new
units. This avoids inflating the protocol state unnecessarily during
periods where too many validators have an outage.