NDRS-868: Clean state synchronously.
For some unknown reason, cleaning linear chain state in the storage
asynchronously (via events) ended up with a node not being able to
participate in the consensus after transitioning to the validator
reactor. I spent few hours debugging this issue but found nothing. Given
the highest priority of this work I decided to go with the alternative.