Add a parent to CandidateBlock.
With deploy replay protection the validity of a candidate block depends
not only on its contents and its timestamp, but also on its ancestors.
So if there are two different vertices with identical consensus values, i.e.
candidate blocks, one of them could be valid and one invalid. Whether a
node would consider them valid would depend on which one was received
first, because that would determine the set of ancestors we check
against.

Adding the parent hash makes candidate block ancestors explicit so that
consensus can rely on the validity result again: If a candidate block is
deemed valid, we know that all vertices that carry it are valid.