Distinguish whether we know that a dependency is satisfied.
A `UnitBySeqNum(s, v, h)` dependency — unit by validator `v` with
sequence number `s` as seen by the unit with hash `h` — can only be
_known_ to be satisfied if we have unit `h` or if we have _no_ unit with
that sequence number by `v`. Otherwise, if we have _some_ unit with that
sequence number, we can't be sure it's the right one. However, for the
purpose of synchronization, we need to treat it as satisfied, so we
continue downloading further dependencies. If it turned out to be wrong,
computation of the panorama will fail and `UnitWithPanorama(h)` will be
considered the next dependency.