Add more asserts to run_equivocator_network.
Assert that the equivocator is not a validator anymore after up to 20
eras. The probability of no equivocation happening should be negligible
in this test, but if it happens, this would tell us why the test failed.

Also assert that the equivocator is only in the `equivocators`
collection, and not e.g. in `inactive_validators`. It shouldn't be
possible for doppelgänger detection to make both nodes appear inactive,
but if that somehow were the case, this would alert us.