Improve and add logging to run_equivocator_network test.
The test is flaky on CI but always passes locally. This adds some
logging so we can see whether it's a timing issue.

We also start isolating one of the two clone nodes for three rounds
starting from the first _consensus_ message it sends now, instead of the
first message it receives: If other messages are exchanged before the genesis
timestamp (address gossiping?), the three round lengths could expire
before the first consensus message is even sent.