Implement test to observe new stack height.
Decrease stack height from 65k to 188.

This mitgates attacks such as naive fibonacci or ackermann's function.
The minimum height that passes the test suite was measured to be 73
which will still give us plenty of room.

AssemblyScript seems to generate wasm that uses longer stack heights in
their runtime.

This value was originally researched to be 128 but we increased it a bit
to still support code generated by assemblyscript compiler v0.9.1.