Make PointerBlock representation denser.
This serializes PointerBlock into a dense structure where first 32 bytes
is the bitmap of occupied Pointer's in the PointerBlock, and then next
(N+7)/8 bytes are dedicated to a bitmap containing map of Leaf/Node
pointers.

Compared to old sparse method we can gain a minimum of 224 bytes saved,
and maximum of 448 bytes per PointerBlock depending on the fill ratio.