Pass a known state root hash from highest block.
This will pass a state root hash coming from a highest block (if exists)
at the initialization time. This is to provide a safe, known snapshot of
the global state.

If a request is made to query an old state which doesn't contain
`Key::SystemContractRegistry` then a fallback query is made using the
"known valid" state.