Inlined I=NodeId: the uninteresting parts
This is the beginning of work to remove the generic I parameter from the
consensus protocol, and thus the application. It's left in an unfinished
state here because much of this was boilerplate of removing <I>s and
changing I to NodeId in argument positions.

In the following commit, the follow-on work to this boilerplate is completed.