Make sync message smaller, use bit fields.
Sync messages now use a bit field for a subset of validators in only a
single round to tell a peer about that part of the sender's protocol
state. However, independent of whether a quorum has been reached, it
always lists the individual validators in that set from which a vote
or echo has been received. Unlike the previous approach this guarantees
that everyone will eventually either see a quorum of "false" votes, or a
quorum of echos, even if the correct validators are not fully connected.