Add `chain_execute_deploy` RPC endpoint.
This new endpoint accepts two parameters:
* block_hash - on top of which to execute the deploy
* deploy - a deploy to execute.

As a result, it will execute the deploy on top of the specified block
and return the results of the execution to the caller. The effects of
the execution won't be committed to the trie (blockchain
database/GlobalState). Endpoint can be used for debugging, discovery -
for example price estimation. If block_hash can't be found, proper error
will be returned. Any error returned during execution will be forwarded
to the caller of the API.

NOTE: The e