Don't add faulty echos/votes after handle_faulty; filter faulty in reply to SyncState.
If we detect double-signing we remove all the echos and votes of the
faulty validator, but we used to still add the second of the two
conflicting signatures afterwards. That was a bug.

We now also don't send echos and votes of faulty validators when
responding to SyncState messages, since these are redundant.