Used extended deploy metadata in JSON-RPC
This commit:
- Changes the storage component's `GetDeployAndMetadata`, which
  is used by the RPC server, to return block information about
  a deploy when execution results are not available.
- Changes the RPC server to use the new `GetDeployAndMetadata`
  for a `GetDeploy` RPC request.
- Updates the JSON schema with the new optional block information
  for a `GetDeployResult`

`StorageRequest::GetDeployAndMetadata` first tries to get the
associated execution results, just like before. If there aren't
any available, instead of returning an empty execution results
`Vec`, it now tries to get the block hash and height of the block
this deploy belongs to. These are stored in the `deploy_hash_index`
where we now also store the block height in addition to the hash.
If the deploy hash wasn't indexed but the deploy was found, we
return empty metadata.

Signed-off-by: George Pisaltu <georgep@casperlabs.io>