Restart current round after unpausing.
This resets `self.current_timeout` after unpausing. The effect is both
that the current round's proposer immediately creates another proposal
(unless they've already proposed), and that the other validators restart
their timeout, so there is a chance the proposal will get accepted.

In nctl tests with block time 0 the pause mode is often activated and
deactivated in quick succession, which caused a lot of proposals getting
dropped. This change fixes that problem.