Improve logging: less spam, more cases.
This changes some log messages to use `Display` instead of `Debug`, and
omits some large data, to make them more readable.

It also adds logging in Zug for when a round becomes skippable, when it
actually gets skipped, and when a round becomes committed but is still
waiting for the accepted proposal.