Avoid redundant update loops; log new rounds.
This adds an info log message when a new round begins, and makes the
`update` function spend only one iteration in each round, resetting the
dirty round to `None` at the end.

Before, it would sometimes mark the round as dirty that it's currently
handling, because its own echoes or votes can create a new quorum.
However, `update_round` processes everything in the right order, so that
any new quorum will be handled immediately.