Remove commented out tests for `get_trie`.
The original test was written against the API of the storage and
recently moved to contract_runtime.rs during extraction of TrieOrChunk.
That test queried trie storage to test that when data gets chunked it
can be restored to its original form later. It set up the trie in such a
way the structure of the trie led to a leaf exceeded the chunk size limit.
After generalizing (and extracting) it to ValueOrChunk and moving
chunking to the constructor, we can test it in isolation without the
special setup of the trie. That test exists in value_or_chunk.rs file.