Replace type bound on V in ValueOrChunk constructor.
We want to chunk when the serialized representation of a type exceeds
our limit. When determining byte representation we should use the same
method that we will use when we actually serialize the data for sending.
Currently, we use bincode::serialize (see Message::new_get_response) so
we should also use it when we chunk.