Use ToBytes impl, instead of bincode, when chunking.
When we write `Trie` to trie store, or when we calculate hash of the
trie, we use `ToBytes` impls to go from the trie to bytes, before
calling [`Digest::hash_bytes_into_chunks_if_necessary`] on the bytes. If
we used different serialization (type to bytes conversion) when
chunking, the proofs in the chunks would not hash to the merkle tree of
the serialized trie from the trie store == different bytes will lead to
different merkle proofs.

This alignment needs to be maintained for all types that want to be
chunked and verified against merkle proofs - the same serialization
mapping needs to be used when calculating merkle root hash and when
chunking.

It seems the easiest to use `bytesrepr::ToBytes` implementations
everywhere, rather than making this generic and endforcing with
scattered, e2e tests.