Always consider highest stored block as trusted block.
Even if a trusted hash is specified in the config file, fast sync will
now use the highest block from storage as a starting point if that is
more recent.

So e.g. a node that started originally with block 1000 as a trusted
block, and was then restarted after block 2000, with the same config
file, will now start syncing from block 2000 forward, not 1000, without
loading the intermediate 1000 headers from storage first.