block_accumulator: don't mark blocks completed
The block accumulator would mark a block completed in storage if it was
executed. Since the block is already marked as completed in the reactor
control logic after a block is executed and announced and also by the
block synchronizer when it's synchronizing historical blocks, there's
no need to mark it complete again in the block accumulator.
This fixes a bug where the block was marked completed two times in rapid
succession.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>