deploy_buffer: fix block continuity check when validating full ttl awareness
When checking if we have full ttl worth of deploys in order to detect
deploy replays, we also check for gaps in the chain index that we have.
Because this check was made after we check the timestamp, it may be
possible to miss a discontinuity if the previous block is not at the
propper height but has a timestamp older than the ttl.
Fix this issue by moving the timestamp check after the block continuity
check.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>