global_state_synchronizer: service next request if another was cancelled due to error
If a request was cancelled due to a fetch error coming from the trie
accumulator, start servicing other requests that may be pending in the
fetch queue. Otherwise we need to wait for a new request to be made or a
trie to be successfully fetched before we can serve another request even
though we have parallel fetch slots available.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>