Use highest block header from storage when refreshing the contract runtime
We may receive block headers that we don't have the body for in a
SyncLeap. Since the header will be added in storage and to the height index,
`read_highest_block` will return `Ok(None)` if it is called just after the
SyncLeap and before getting the body. This will prevent the contract
runtime to be correctly initialized, making it stall execution of
subsequent blocks.
Fix this by using the highest block header instead of the highest block
from storage while refreshing the contract runtime.

Co-authored-by: George Pisaltu <georgep@casperlabs.io>
Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>