block_synchronizer/peer_list: change Unknown peer behaviour
The only time a peer has `Unknown` quality is when it was just added to
the peer list and we haven't heard anything from it.
If a peer that has `Unknown` quality is promoted we now make it
`Reliable` since it means that an item we required was found on that
node.
Also if a peer with `Unknown` quality is demoted, we make it
`Unreliable`.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>