block_synchronizer: propagate unreliable peers for global state sync
The trie accumulator uses a snapshot of the peer list provided by the
block synchronizer through the global state synchronizer in order to get
the required tries.
Propagate the unreliable peers (that didn't have the requested data)
upwards to the global state synchronizer and then to the block
synchronizer. This gives feedback that the original peer list provided
should be changed if another global sync request is made.
Demote the peers received from the global state synchronizer in the
block synchronizer.
Also add a test to verify that the peer list is changed based on the
feedback from the global state sync components.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>