Introduce `UnexecutedBlockAnnouncement`
This commit adds a new announcement for finalized blocks enqueued for
execution, but which contract runtime will not execute because their
height is lower than the next expected executable height. In this
scenario, contract runtime will raise this announcement which will be
handled by the reactor. If there is a complete block in storage with
that same height, it means the block can be safely considered executed
and the reactor raises an event letting consensus know that the
finalized block it had enqueued was handled, the same way it would
happen if the block was executed. Consensus is oblivious to the fact
that execution didn't take place.

We do this to ensure the smooth operation of the consensus validating
process in the case of a restart in the middle of an era, where the
restarted node couldn't make any consensus proposals until all the
previous blocks in the era along with their vertices were "executed".

Signed-off-by: George Pisaltu <georgep@casperlabs.io>