historical sync: fix leap to era 0 of pre 1.5 network
In a pre 1.5 network, we can't get the trusted block headers from era 0
because the sync leap is expected to return the switch block of the
previous era in the sync leap response but it won't if the trusted block
is part of era 0 of a pre 1.5 network where we don't have an immediate
switch block after genesis.

Allow a sync leap response to return the blocks in era 0 if genesis is
in this era.
Also use the method to infer era 0 validators when registering era 1.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>