nctl/nightly tests: fix upgrade scenario 9
We allow a node of the same binary to sync to a trusted hash of an older
protocol version only if the trusted hash is within the allowed sync
leap range. Adjust the test to check that if the trusted hash is not
within the allowed range we don't allow the new node to sync.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>