ContractRuntime: delay init of system contract registry cache
We currently cache the system contract registry in the ContractRuntime
component early in the initialization process although it is not needed
until later. This is a problem when upgrading from a version that did
not have the system contract registry written in the global state.
Allow the ContractRuntime to skip initialization of the system contract
registry until it is actually needed. This allows the upgrade to create
the registry if required and initialize the cache afterwards.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>