block_synchronizer: add a time threshold for detecting stalled builders
Because there are situations where the block synchronizer can get
stuck, added a new config parameter to compare against to check if the
synchronizer didn’t make any progress.
So if a builder did not receive data successfully and updated its
progress for the configured amount of time, we flag it as stalled and
the control logic should purge it and try again.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>