synchronizer: allow progress for pre-1.5 blocks without weak finality
Allow sync to progress for historical blocks that were created with a
protocol version that is older than the one where the requirement for
nodes to hold signatures for strict finality was introduced. In mainnet
this is protocol version 1.5. For other networks a configurable
chainspec parameter was added to control the version when this
requirement was introduced.
Because pre-1.5 there was no requirement for nodes to retain finality
signatures, there may exist networks were all finality signatures for a
block are not available in order to reach weak or strict finality.
Because we have 2 checkpoints for determining if we can progress with
the sync (first achieving weak finality, and then strict finality) we
employ the `LegacyRequiredFinality` the requirement for the block that
is within the criteria described above. Previous code did not take into
account the check for weak finality.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>