Cherry-picked wasmi upgrade to 0.13.2
This wasmi version deprecates allocating `MemoryInstance::get` that
always allocates and copies data back and forth between VM linear memory
and the heap on the host.

With this commit host does not allocate data, but is deserializing data
straight from a linear memory without allocating. Copies are made only
when absolutely necessary.

There's also update of casper-wasm-utils crate in this commit.