fast sync: Fix stale sync leaper activity that affects recovery code
Sometimes the sync leaper leap activity is not correctly reset and
causes the node to fail to recover later when it gets out of sync.
This happens due to a race condition that appears when we update the
validator matrix with the validators for a block after an upgrade.

This happens because the validator matrix update is delayed because
the validators need to be read from global state but by doing so the
update can interleave within the normal determination of the keep up
sync instruction. If the update comes after a leap request is sent,
the leap activity will not be reset since the keep up look may not
need to check the leap status again if it starts to sync a block and
then finishes historical sync.

This patch purges any potential stale sync back leap activities after
the node is done syncing back since any new received leap activity is
not needed anymore (and it will actually never be checked again until
the node switches to CatchUp).

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>