main_reactor: fix determination of `should_commit_upgrade`
The control logic was using the cached `switch_block_header` to
determine if the node is ready to commit an upgrade.
This field was not only updated with the header of the latest complete
block and was set also by the `CatchUp` logic to the highest block
header that the node has stored. In the case where a `SyncLeap` is
received, the headers in that message are stored. This in turn can lead
`switch_block_header` to hold a header of a block that may not have been
marked as completed.

Because of this the control logic was tricked into thinking that it
should not commit an upgrade (because there was a header that was of the
newer protocol version received through SyncLeap) and switch to `KeepUp`
and then to `Validate` (if it was supposed to be a validator) even if
the node did not have the immediate switch block marked as complete.
This can happen if the node that is supposed to be validating comes up
after an upgrade a little later than the other nodes in the network; in
this case it comes late enough such that its peers already completed and
are responding to a SyncLeap with the newly created immediate switch
block.

Co-authored-by: George Pisaltu <georgep@casperlabs.io>
Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>