main_reactor: don't allow upgrade to loop indefinitely
In some cases, the node can become stuck trying to commit the upgrade
due to different reasons. One scenario is that of a node that comes up
a little later than the others after an emrgency upgrade and tries to
commit the upgrade. Because the other validators may have already
finished gossiping the new immediate switch block, this node may not
have heard the gossip since it was down. In this situation the node will
not be able to store the immediate switch block it created since it
can't hear about sufficient signatures from the other peers.
Instead of looping in the `Upgrading` state indefinitely without no
chance of being able to store the block, if enough time has passed we
switch the node back to `CatchUp`. In this state it can determine if it
is at tip or not and either try to upgrade again or sync back the
immediate switch block from the network.

Co-authored-by: George Pisaltu <georgep@casperlabs.io>
Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>