sync_leap: don't filter all switch blocks preceding upgrades
We can determine if an upgrade did change the validator set or not by
looking at the chainspec. If the upgrade to the current protocol version
did not alter the validator set, then we can use the validators in the
switch block preceding the upgrade.
This commit relaxes the filtering of switch blocks before upgrades to
only those for which we can't reliably determine if a change in
validators happened or not by looking at the chainspec.

Co-authored-by: George Pisaltu <georgep@casperlabs.io>
Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>