ee/tests: extend tests to check V deserialization on read/write
Extend testing logic to check if reading a value from a trie
deserializes it once if the key is found, or zero times if the key does
not exist.
Extend testing logic to check if writing a value doesn't deserialize it
during a write operation.
Extended the existing counter that tracks calls to `from_bytes` for `V`
to work with `read` and `write` operations.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>