consensus: don't error when deactivating era more than once
Don't return a fatal error when trying to deactivate an era that was
already deactivated.
In the case where a node unbonds and drops from the validator list, it
will switch to `KeepUp` in order to follow the chain. Switching to
`KeepUp` will try to deactivate the current era which would have been
already deactivated when the last block of the era was added.
Print out a debug message to track when this happens.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>