nctl: test scenario for sync to genesis after restart when sync in era 0
Before 1.5.0 there wasn't an immediate switch block committed at
genesis. Because historical sync relies on sync leaps to get the
validators for a particular era, we encounter a special case
when syncing the blocks of era 0 if they were created before 1.5.0
because sync leap will not be able to include a switch block that
has the validators for era 0.
Since the auction delay is at least 1, we can generally rely on
the fact that the validator set for era 1 and era 0 are the same.

Add a test to check if a node that has synced back to some block in
era 0 can continue syncing to genesis after it was restarted (and
lost its validator matrix).

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>