types/block: remove structs used for JSON representation
We can now use the `serde-map-to-array` crate to generate schemas for
map type objects using `schemars` directly. Since we used `JsonBlock*`
structs before because we had this limitation, this commit removes these
structs and derive `JsonSchema` for `Block` and associated structs.
We also fix up the responses for RPC, event stream and REST servers so
that they don't perform a conversion to BlockV2 from a versioned block.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>