versioned block: don't convert down to `BlockV2`
Converting from a versioned block to `BlockV2` is not always supported
if the versioned block is not of the `V2` variant.
Implement `TryFrom` instead of `From` to enforce this.

Also the block accumulator requires blocks of the current version
because it should only care about recent blocks that are being gossiped
and not old historical blocks. Currently when a block is fetched it is
also sent to the block accumulator. If we can't convert the fetched
block to a current version it means that the block is historical and
there's no need to send it to the accumulator.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>