nctl: fix upgrade_after_emergency_upgrade_test_pre_1.5
In some runs this upgrade test was failing because it was hitting the
300 second timeout for syncing to genesis.
Sometimes when synchronizing historical blocks, the block synchronizer
for node 1 was selecting 5 other nodes that did not have the finality
signatures for that block (usually nodes 6-10).
Because the peer refresh interval was set to 90 seconds, the node did
not have time to finish fetching the historical blocks if the situation
described above happened for multiple blocks since the total timeout was
300 seconds.

Reduce the `block_synchonizer.peer_refresh_interval` config parameter
for this test to 10s to allow it to become more lively and finish within
the 300s timeout.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>