Fix allocation in `FromBytes for Vec<T>`
Deserialization of arbitrary vectors could end up in different outcomes
depending on the OS and the configuration. This can be then exploited by
crafting a Wasm that will force host to preallocate 2^32-1 * size_of<T>
bytes which will fail (with either OutOfMemory or EarlyEndOfStream) and
then this information can get in a block hash.

This commit fixes this behavior and will try to preallocate as little as
possible and will not rely on system's allocator behavior.