casper-storage: add basic block store functionality from the node
Initial attempt to add some of the basic CRUD block store operations
to casper-storage. This allows other crates to interact with the node
storage databases without the need to go through the storage
component of the node.

The `block_store` module in casper-storage now provides traits to
interact with a casper block store. It also provides concrete
implementations for using LMDB as the backing store and also allows
accessing indexed data.

Some types that were defined in the node have moved to casper-types and
casper-storage as a result of this implementation.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>