Try to emulate wasm traps under x86
This is (currently) aimed at `casper_return` which we have somehow
regain control after its executed, and capture the output.

When casper_return is called it sets a thread local value which can be
later inspected. The downside is that we can't use ! on x86 as it
causes seg faults - probably, because the symbol assumes it's return
type is ! but the code executes fine.

It's ugly but it seems like does the work and at this stage it should
help with development of the prototype.