Support for stateful and private trait methods.
Currently all entrypoints in a trait are dispatched and state is always
written, but it's a subject to refactor to match behavior of an impl
block.

Private trait methods can be used to implement providers i.e. a trait
can expose a "State" struct that has all information needed to serve a
logic implemented in a trait.