ee/contract messages: add global per block index for contract messages
In order to be able to track the ordering of messages that are emitted
by different contracts, we now generate an unique ordering number for
each message that is emitted during the execution of a block.

The first message that is emitted when a block is executed has
`block_index=0`, the second message has `block_index=2` and so on.
We also write a record with the total number of messages that were
emitted during the execution of a block. This record is purged from
global state if a new message is emitted in a new block.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>