There is no need to cast a contract handle
With this commit each trait gets a "Ext" trait as well that returns
ToCallData structures, and a Ref struct for that trait that implements this trait.

This way with the help of a macro we can impl Ext structs for each
`impl_traits` types, and magically a #[casper(contract)] struct gets
access to all trait methods to build calls.