Makefile: fix rust smart contracts build
The contracts built using `make build-contracts-rs` were not working on
a production network because they included the `casper_print` FFI. This
happened because of the way `cargo` optimizes builds of dependencies.
Because one of the contracts had a requirement for the `test-support`
feature, all contracts were built with the `casper-contract` that had
that feature enabled (and which enabled the `casper_print` FFI).
Now we build each smart contract idependently to make sure that never
happens.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>