chainspec/contract_messages: update costs for `emit_message` parameters
Because `emit_message` and `manage_message_topic` perform operations in
the backend that may be expensive (e.g. hashing a topic name or the
message payload), we set non-zero default values for costs related to
parameters involved in these types of operations. For example emitting a
message of a small payload size (e.g. 10 bytes) is cheaper than emitting
a message with larger payload (e.g. 1000 bytes) since these bytes need
to be hashed and checked.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>