scratch: optimise `keys_with_prefix` to allow faster lookup of cached keys
Use a trie to store all the keys that are cached. This allows for
efficient lookup of all the keys starting with a provided prefix. Up
until now we were iterating thorugh all the keys in the cache to check
if they have the specified prefix which was a constly operation when a
lot of values were cached.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>