Fix issue with multiple balance holds in the same block.
When multiple transactions from the same initiator are included in the
same block, balance holds created for gas should accumulate for all the
transactions in that block.
We were previously overwriting the gas hold record in global state with
the hold for the last transaction executed in the block for the
corresponding purse.
We now add the hold amount in the single record.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>