trie_store: add batch write operation
When inserting a batch of values to the trie store we don't really care
about interstitial state root hashes. Writing (key, value) pairs to the
trie store sequentially will cause the trie to re-hash after every
insert. This leads to poor insert performace and also increased storage
space to hold trie nodes that will potentially never be accessed.

In this PR we create an in-memory cache of the trie backed by the
store. Existing nodes are read back from the trie store when needed.
After the entire batch is written in memory, we traverse the trie once
to convert the in-mem nodes, hash and store them to disk, resulting in a
single state root hash for the entire batch write operation.

The cache insert algorithm matches the sequential write algorithm such
that writing (key, value) pairs sequentially produces the exact same
end state trie (and of course the exact same state root hash) as
inserting pairs using a batch write operation.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>